Basic elements of the Home Page
I added two navbars, in the first one there is the 'Logo" and two options for register and login. in the other one we have three more links attached that are home, about us and services. Underneath the navbar i attached a image of a view that is related to the purpose of our platform. I plan on adding more content.